refactor: moved duplicated code into hook
